+++ R/targets.R
@@ -4,6 +4,10 @@
 #' obtained from Ruijter (2013), \doi{10.1016/j.ymeth.2012.08.011} but original
 #' source is by Vermeulen (2009), \doi{10.1016/S1470-2045(09)70154-8}.
 #'
+#' The tidy version of the data is kept at the repository of the source of
+#' `{vermeulen}` package. This function fetches such data and thus requires
+#' internet connection.
+#'
 #' @return A data frame with 64 targets and two variables:
 #'
 #' \describe{
@@ -15,8 +19,11 @@
 #' @seealso `amplification_curves()` `reactions()` `samples()`
 #'
 #' @source
-#' - Vermeulen (2009), \doi{10.1016/S1470-2045(09)70154-8}.
-#' - Ruijter (2013), \doi{10.1016/j.ymeth.2012.08.011}.
+#' - \url{https://medischebiologie.nl/wp-content/uploads/2019/02/qpcrdatamethods.zip}
+#' - \url{https://github.com/ramiromagno/vermeulen/blob/main/data-raw/targets.csv.gz}
+#'
+#' @examples
+#' try(targets())
 #'
 #' @export
 targets <- function() {