+++ NEWS.md
@@ -1,3 +1,14 @@
+# vermeulen 0.1.1
+
+* A new function `ds_biomarker()` allows to download the entire Biomarker data
+set with one call.
+
+* Fetching functions are now memoised thus providing greater convenience when
+running interactively.
+
+* The order of the levels of the column `well` is now fixed: A1, A2, ... instead
+of A1, A10, ...
+
 # vermeulen 0.1.0
 
 * First release of the package.